@@ -3,6 +3,7 @@
   "version": "0.3.1",
   "description": "MapSCII is a Braille & ASCII world map renderer for your console, based on OpenStreetMap",
   "main": "main.js",
+  "type": "module",
   "scripts": {
     "lint": "eslint src",
     "start": "node main",
@@ -16,7 +17,7 @@
     "mapscii": "./bin/mapscii.sh"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=20"
   },
   "keywords": [
     "map",
@@ -42,11 +43,12 @@
     "string-width": "^4.2.0",
     "term-mouse": "^0.2.2",
     "x256": "0.0.2",
-    "yargs": "^15.4.1"
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
+    "@babel/preset-typescript": "^7.25.7",
     "eslint": "^7.8.1",
     "eslint-plugin-jest": "^24.0.0",
-    "jest": "^26.4.2"
+    "jest": "^29.7.0"
   }
 }
