@@ -6,32 +6,39 @@
   * remote TileServer
   * local MBTiles and VectorTiles
 */
-'use strict';
-const fs = require('fs');
-const path = require('path');
-const fetch = require('node-fetch');
-const envPaths = require('env-paths');
+import fs from 'fs';
+import path from 'path';
+import fetch from 'node-fetch';
+import envPaths from 'env-paths';
 const paths = envPaths('mapscii');
 
-const Tile = require('./Tile');
-const config = require('./config');
+import Tile from './Tile';
+import config from './config';
 
 // https://github.com/mapbox/node-mbtiles has native build dependencies (sqlite3)
 // To maximize MapSCII’s compatibility, MBTiles support must be manually added via
 // $> npm install -g @mapbox/mbtiles
 let MBTiles = null;
 try {
-  MBTiles = require('@mapbox/mbtiles');
+  MBTiles = await import('@mapbox/mbtiles');
 } catch (err) {void 0;}
 
-const modes = {
-  MBTiles: 1,
-  VectorTile: 2,
-  HTTP: 3,
+export enum Mode {
+  MBTiles = 1,
+  VectorTile = 2,
+  HTTP = 3,
 };
 
 class TileSource {
-  init(source) {
+  private source: string;
+  private cache: any;
+  private cacheSize: number;
+  private cached: any;
+  private mode: Mode | null;
+  private mbtiles: any;
+  private styler: any;
+
+  init(source: string) {
     this.source = source;
     
     this.cache = {};
@@ -47,14 +54,14 @@ class TileSource {
         this._initPersistence();
       }
 
-      this.mode = modes.HTTP;
+      this.mode = Mode.HTTP;
 
     } else if (this.source.endsWith('.mbtiles')) {
       if (!MBTiles) {
         throw new Error('MBTiles support must be installed with following command: \'npm install -g @mapbox/mbtiles\'');
       }
 
-      this.mode = modes.MBTiles;
+      this.mode = Mode.MBTiles;
       this.loadMBTiles(source);
     } else {
       throw new Error('source type isn\'t supported yet');
@@ -77,7 +84,7 @@ class TileSource {
     this.styler = styler;
   }
 
-  getTile(z, x, y) {
+  getTile(z: number, x: number, y: number) {
     if (!this.mode) {
       throw new Error('no TileSource defined');
     }
@@ -95,14 +102,14 @@ class TileSource {
     }
   
     switch (this.mode) {
-      case modes.MBTiles:
+      case Mode.MBTiles:
         return this._getMBTile(z, x, y);
-      case modes.HTTP:
+      case Mode.HTTP:
         return this._getHTTP(z, x, y);
     }
   }
 
-  _getHTTP(z, x, y) {
+  private _getHTTP(z: number, x: number, y: number) {
     let promise;
     const persistedTile = this._getPersited(z, x, y);
     if (config.persistDownloadedTiles && persistedTile) {
@@ -122,7 +129,7 @@ class TileSource {
     });
   }
 
-  _getMBTile(z, x, y) {
+  private _getMBTile(z: number, x: number, y: number) {
     return new Promise((resolve, reject) => {
       this.mbtiles.getTile(z, x, y, (err, buffer) => {
         if (err) {
@@ -133,7 +140,7 @@ class TileSource {
     });
   }
 
-  _createTile(z, x, y, buffer) {
+  private _createTile(z: number, x: number, y: number, buffer) {
     const name = [z, x, y].join('-');
     this.cached.push(name);
     
@@ -141,7 +148,7 @@ class TileSource {
     return tile.load(buffer);
   }
 
-  _initPersistence() {
+  private _initPersistence() {
     try {
       this._createFolder(paths.cache);
     } catch (error) {
@@ -149,14 +156,14 @@ class TileSource {
     }
   }
 
-  _persistTile(z, x, y, buffer) {
+  private _persistTile(z, x, y, buffer) {
     const zoom = z.toString();
     this._createFolder(path.join(paths.cache, zoom));
     const filePath = path.join(paths.cache, zoom, `${x}-${y}.pbf`);
     return fs.writeFile(filePath, buffer, () => null);
   }
 
-  _getPersited(z, x, y) {
+  private _getPersited(z: number, x: number, y: number) {
     try {
       return fs.readFileSync(path.join(paths.cache, z.toString(), `${x}-${y}.pbf`));
     } catch (error) {
@@ -164,7 +171,7 @@ class TileSource {
     }
   }
 
-  _createFolder(path) {
+  private _createFolder(path: string): true {
     try {
       fs.mkdirSync(path);
       return true;
@@ -175,4 +182,4 @@ class TileSource {
   }
 }
 
-module.exports = TileSource;
+export default TileSource;
